amelioration de l'affichage
